moving things around and cleanup
