+++ docker-compose.yml
@@ -5,6 +5,7 @@ services:
       context: .
     ports:
       - '80:3000'
+      - '88:8888'
       - '443:4443'
       - '7003:7003'
       - '8226:8226'
