+++ src/WebServer.ts
@@ -14,25 +14,91 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+import * as fs from "fs";
+import { IncomingMessage, ServerResponse } from "http";
+import * as https from "https";
+import { SSLConfig } from "./ssl-config";
+
 import { IServerConfiguration } from "./IServerConfiguration";
 import { logger } from "./logger";
-import { PatchServer } from "./patchServer";
-import { WebServer } from "./web";
 
-export class Web {
-  /**
-   * Start HTTP and HTTPs connection listeners
-   * TODO: This code may be better suited in web.js and patchServer.js
-   */
-  public async start(config: IServerConfiguration) {
-    // Start the mock patch server
-    const patchServer = new PatchServer();
-    patchServer.start(config);
-    logger.info("[webServer] Patch Server started");
+/**
+ * Create the SSL options object
+ */
+function sslOptions(configuration: IServerConfiguration["serverConfig"]) {
+  const sslConfig = new SSLConfig();
+  return {
+    cert: fs.readFileSync(configuration.certFilename),
+    ciphers: sslConfig.ciphers,
+    honorCipherOrder: true,
+    key: fs.readFileSync(configuration.privateKeyFilename),
+    rejectUnauthorized: false,
+    secureOptions: sslConfig.minimumTLSVersion,
+  };
+}
 
-    // Start the AuthLogin and shardlist servers
-    const webServer = new WebServer();
-    webServer.start(config.serverConfig);
-    logger.info("[webServer] Web Server started");
+function httpsHandler(
+  request: IncomingMessage,
+  response: ServerResponse,
+  config: IServerConfiguration["serverConfig"]
+) {
+  logger.info(
+    `[HTTPS] Request from ${request.socket.remoteAddress} for ${
+      request.method
+    } ${request.url}`
+  );
+  switch (request.url) {
+    case "/cert":
+      response.setHeader(
+        "Content-disposition",
+        "attachment; filename=cert.pem"
+      );
+      response.end(fs.readFileSync(config.certFilename));
+      break;
+
+    case "/key":
+      response.setHeader("Content-disposition", "attachment; filename=pub.key");
+      response.end(fs.readFileSync(config.publicKeyFilename).toString("hex"));
+      break;
+    case "/AuthLogin":
+      response.setHeader("Content-Type", "text/plain");
+      response.end("Valid=TRUE\nTicket=d316cd2dd6bf870893dfbaaf17f965884e");
+      break;
+
+    default:
+      if (request.url && request.url.startsWith("/AuthLogin?")) {
+        response.setHeader("Content-Type", "text/plain");
+        response.end("Valid=TRUE\nTicket=d316cd2dd6bf870893dfbaaf17f965884e");
+        return;
+      }
+      response.statusCode = 404;
+      response.end("Unknown request.");
+      break;
   }
 }
+
+export class WebServer {
+  public async start(config: IServerConfiguration["serverConfig"]) {
+    const httpsServer = https
+      .createServer(
+        sslOptions(config),
+        (req: IncomingMessage, res: ServerResponse) => {
+          httpsHandler(req, res, config);
+        }
+      )
+      .listen({ port: 4443, host: "0.0.0.0" })
+      .on("connection", socket => {
+        socket.on("error", (error: Error) => {
+          logger.error(`[webServer] SSL Socket Error: ${error.message}`);
+        });
+        socket.on("close", () => {
+          logger.info("[webServer] SSL Socket Connection closed");
+        });
+      })
+      .on("tlsClientError", err => {
+        logger.error(`[webServer] tlsClientError: ${err}`);
+      });
+  }
+}
+
+module.exports = { WebServer };
