+++ src/packet.test.js
@@ -14,22 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-const packet = require('./packet');
+const packet = require("./packet");
 
-test('retuns a buffer packet from premadeLogin', () => {
+test("retuns a buffer packet from premadeLogin", () => {
   expect(packet.premadeLogin()).toBeInstanceOf(Buffer);
-});
-
-test('retuns a buffer packet from craftGenericReply', () => {
-  expect(packet.craftGenericReply()).toBeInstanceOf(Buffer);
-});
-
-test('retuns a buffer packet from premadePersonaMaps', () => {
-  expect(packet.premadePersonaMaps()).toBeInstanceOf(Buffer);
-});
-
-test('retuns a buffer packet from buildPacket', () => {
-  expect(
-    packet.buildPacket(6, 0x601, Buffer.from([0x06, 0x01])),
-  ).toBeInstanceOf(Buffer);
 });
