+++ src/packet.ts
@@ -14,19 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-export function buildPacket(len: number, header: number, content: Buffer) {
-  // TODO: Confirm that len is equal to the content len
-  const packet = Buffer.alloc(len);
-
-  // Add the response code
-  packet.writeUInt16BE(header, 0);
-
-  // Write the content
-  content.copy(packet, 2);
-
-  return packet;
-}
-
 /**
  * This is the response packet sent on the login port in response to a UserLogin
  * TODO: Replace the need for this
@@ -290,1071 +277,5 @@ export function premadeLogin() {
     0x00, // (0x8c was old val)
     0x00, // (0x98 was old val)
     0x00, // (0xa6 was old val)
-  ]);
-}
-
-export function craftGenericReply() {
-  const packet = Buffer.from([
-    0x31,
-    0x00,
-    0x54,
-    0x4f,
-    0x4d,
-    0x43,
-    0x01,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x65,
-    0x00,
-    0xb6,
-    0x01,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-  ]);
-
-  return packet;
-}
-
-// TODO: Remove the need for this
-export function premadePersonaMaps() {
-  return Buffer.from([
-    // Live
-
-    0x72,
-    0x7e,
-    0xe2,
-    0xe2,
-    0xcd,
-    0x30,
-    0x75,
-    0x9a,
-    0x90,
-    0x37,
-    0x4d,
-    0xd0,
-    0x37,
-    0x77,
-    0x02,
-    0xea,
-    0x9b,
-    0x20,
-    0x19,
-    0x24,
-    0xa7,
-    0xbe,
-    0x57,
-    0x37,
-    0xf4,
-    0x32,
-    0x6c,
-    0x78,
-    0x0a,
-    0x0d,
-    0x51,
-    0xad,
-    0x81,
-    0x0b,
-    0xfe,
-    0x07,
-    0x5a,
-    0x71,
-    0x88,
-    0xc1,
-    0xdd,
-    0x92,
-    0x08,
-    0x42,
-    0x81,
-    0xd2,
-    0x08,
-    0xdb,
-    0x5e,
-    0x08,
-    0xb7,
-    0x93,
-    0x09,
-    0x50,
-    0xb9,
-    0x2e,
-    0x13,
-    0xf6,
-    0xcf,
-    0xe3,
-    0x3c,
-    0xd8,
-    0x23,
-    0xb5,
-    0xa4,
-    0xc9,
-    0x06,
-    0x29,
-    0x6b,
-    0xd7,
-    0xc9,
-    0xcb,
-    0xb1,
-    0x7a,
-    0x80,
-    0x0a,
-    0xff,
-    0x81,
-    0xd8,
-    0x43,
-    0x88,
-    0xed,
-    0x91,
-    0x69,
-    0xf0,
-    0xdc,
-    0x10,
-    0xc5,
-    0x4b,
-    0x01,
-    0x2f,
-    0x8e,
-    0xd2,
-    0xe0,
-    0x7f,
-    0x3f,
-    0x58,
-    0x25,
-    0x26,
-    0xb9,
-    0x4e,
-    0x55,
-    0xd4,
-    0x2c,
-    0xeb,
-    0xb6,
-    0x37,
-    0x2f,
-    0x80,
-    0x22,
-    0xd6,
-    0x93,
-    0x9a,
-    0xd7,
-    0xbf,
-    0x28,
-    0x63,
-    0x8f,
-    0xca,
-    0x3d,
-    0x15,
-    0x68,
-    0xeb,
-    0x84,
-    0x97,
-    0x0b,
-    0x62,
-    0x8d,
-    0x4f,
-    0x0a,
-    0x2a,
-    0xf0,
-    0x0e,
-    0x3a,
-    0x9d,
-    0xae,
-    0x24,
-    0x60,
-    0x06,
-    0x6b,
-    0x30,
-    0x55,
-    0x55,
-    0x11,
-    0x99,
-    0x8e,
-    0xd2,
-    0xcd,
-    0xc9,
-    0xcc,
-    0x70,
-    0xb0,
-    0xa3,
-    0x43,
-    0xf8,
-    0x7d,
-    0x83,
-    0x26,
-    0x19,
-    0xb1,
-    0x06,
-    0x07,
-    0x01,
-    0x01,
-    0x73,
-    0x92,
-    0x01,
-    0xfe,
-    0x74,
-    0x89,
-    0x52,
-    0x58,
-    0x00,
-    0x01,
-    0x11,
-    0x6d,
-    0x28,
-    0xd7,
-    0x9f,
-    0xe9,
-    0x00,
-    0x00,
-    0x00,
-    0x01,
-    0x00,
-    0x00,
-    0x00,
-    0x2c,
-    0x78,
-    0xaa,
-    0xd0,
-    0x10,
-    0xad,
-    0x4f,
-    0x44,
-    0x6f,
-    0x63,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0xc0,
-    0xa0,
-    0x62,
-    0x38,
-    0xe8,
-    0xe0,
-    0xf3,
-    0x53,
-    0x97,
-    0x39,
-    0x51,
-    0xe4,
-    0x66,
-    0x01,
-    0xbd,
-    0xd3,
-    0xa6,
-    0x74,
-    0x2d,
-    0x96,
-    0x00,
-    0xb7,
-    0x16,
-    0x35,
-    0x57,
-    0xa5,
-    0x10,
-    0xb4,
-    0x0d,
-    0x7d,
-    0xf5,
-    0xa9,
-    0x05,
-    0xeb,
-    0x23,
-    0x5d,
-    0x03,
-    0xbc,
-    0x95,
-    0x11,
-    0xeb,
-    0xb2,
-    0x1b,
-    0xdf,
-    0x26,
-    0x16,
-    0xd7,
-    0xb3,
-    0x1e,
-    0xdf,
-    0x63,
-    0x0f,
-    0x6b,
-    0xc9,
-    0x32,
-    0x05,
-    0x75,
-    0x51,
-    0x6e,
-    0x27,
-    0x06,
-    0xcb,
-    0x5d,
-    0x69,
-    0x88,
-    0xa3,
-    0x11,
-    0xf0,
-    0x29,
-    0x08,
-    0x29,
-    0x8e,
-    0xd9,
-    0xb6,
-    0xc8,
-    0x56,
-    0xee,
-    0xff,
-    0x12,
-    0xdf,
-    0xe3,
-    0x33,
-    0xf7,
-    0x34,
-    0x89,
-    0x6a,
-    0xdc,
-    0xd0,
-    0x54,
-    0xc8,
-    0x34,
-    0x69,
-    0xf5,
-    0xa4,
-    0x81,
-    0xa3,
-    0x16,
-    0xff,
-    0xd3,
-    0x4b,
-    0xb9,
-    0xc6,
-    0x52,
-    0x34,
-    0x51,
-    0xa6,
-    0x40,
-    0xcc,
-    0x86,
-    0x1b,
-    0x21,
-    0x56,
-    0x28,
-    0xcf,
-    0xed,
-    0x53,
-    0x82,
-    0x5c,
-    0x55,
-    0xbf,
-    0xed,
-    0xad,
-    0x57,
-    0x51,
-    0x0a,
-    0xea,
-    0x52,
-    0xf6,
-    0x8d,
-    0x12,
-    0xd7,
-    0xad,
-    0x1b,
-    0x7e,
-    0xcf,
-    0x6d,
-    0xec,
-    0x24,
-    0xa5,
-    0x9a,
-    0xca,
-    0x99,
-    0x7e,
-    0x20,
-    0x80,
-    0xa1,
-    0x3b,
-    0x61,
-    0x61,
-    0x0c,
-    0xb6,
-    0xe1,
-    0xbe,
-    0x03,
-    0x5c,
-    0x00,
-    0x02,
-    0xcb,
-    0xb5,
-    0xb6,
-    0x88,
-    0xdc,
-    0xf4,
-    0x73,
-    0xf7,
-    0xce,
-    0xe3,
-    0x48,
-    0xa3,
-    0xa6,
-    0x87,
-    0x5d,
-    0x24,
-    0x1c,
-    0x8a,
-    0x0a,
-    0x60,
-    0x9a,
-    0x02,
-    0xb4,
-    0x71,
-    0xf4,
-    0x10,
-    0x8a,
-    0x24,
-    0xed,
-    0x65,
-    0x4c,
-    0x6d,
-    0xf7,
-    0x29,
-    0x7f,
-    0xbe,
-    0x72,
-    0x20,
-    0xf7,
-    0xbb,
-    0x92,
-    0xf7,
-    0xaa,
-    0xcc,
-    0x14,
-    0xb8,
-    0x90,
-    0x17,
-    0x56,
-    0x67,
-    0x79,
-    0x88,
-    0x08,
-    0x90,
-    0x13,
-    0x5f,
-    0x07,
-    0x40,
-    0x25,
-    0xd3,
-    0xce,
-    0x9e,
-    0x53,
-    0xd1,
-    0xb7,
-    0x5a,
-    0x2b,
-    0xca,
-    0xed,
-    0x9d,
-    0xd9,
-    0x11,
-    0x64,
-    0xc4,
-    0xa9,
-    0x0e,
-    0xfc,
-    0xe5,
-    0x3c,
-    0x71,
-    0x1e,
-    0x79,
-    0x0f,
-    0xf9,
-    0x55,
-    0xab,
-    0x64,
-    0xc9,
-    0x72,
-    0x6a,
-    0xd9,
-    0x85,
-    0x21,
-    0x7f,
-    0x85,
-    0x66,
-    0x15,
-    0x66,
-    0xbc,
-    0xf6,
-    0x79,
-    0x78,
-    0x89,
-    0xcd,
-    0x4f,
-    0x3f,
-    0x4a,
-    0xc3,
-    0x3e,
-    0x90,
-    0xf5,
-    0xc7,
-    0xfd,
-    0xe8,
-    0x46,
-    0x2b,
-    0x76,
-    0x64,
-    0x0e,
-    0xd9,
-    0x50,
-    0x7a,
-    0xdc,
-    0x5d,
-    0x7c,
-    0x91,
-    0x92,
-    0x4c,
-    0xe8,
-    0x6b,
-    0xdf,
-    0xd9,
-    0xc0,
-    0x08,
-    0xfa,
-    0x00,
-    0xb8,
-    0xd4,
-    0xd8,
-    0x20,
-    0xa5,
-    0x2e,
-    0x16,
-    0x0a,
-    0xf8,
-    0x0e,
-    0x0c,
-    0x0a,
-    0xd2,
-    0xdb,
-    0x1e,
-    0xcd,
-    0x57,
-    0x1f,
-    0x35,
-    0x1a,
-    0x9e,
-    0x3c,
-    0x9d,
-    0x94,
-    0xdb,
-    0x97,
-    0x49,
-    0xd2,
-    0x35,
-    0x6e,
-    0xeb,
-    0xf9,
-    0xfc,
-    0xe6,
-    0x70,
-    0x3a,
-    0x22,
-    0xe6,
-    0xc2,
-    0x86,
-    0x99,
-    0x6f,
-    0x3e,
-    0x19,
-    0x5b,
-    0x55,
-    0xe3,
-    0x58,
-    0xe1,
-    0xd1,
-    0x9c,
-    0xb3,
-    0x8e,
-    0x80,
-    0xb3,
-    0x87,
-    0xdb,
-    0x2c,
-    0xf8,
-    0xfb,
-    0x94,
-    0xee,
-    0xa7,
-    0x69,
-    0x88,
-    0xcd,
-    0xfa,
-    0xde,
-    0x1d,
-    0x7c,
-    0xd0,
-    0xae,
-    0x5d,
-    0xdf,
-    0x25,
-    0xf5,
-    0x08,
-    0x77,
-    0x09,
-    0xdb,
-    0xb8,
-    0xee,
-    0x4b,
-    0x40,
-    0xac,
-    0x41,
-    0x57,
-    0x7b,
-    0x0d,
-    0xba,
-    0x6d,
-    0x16,
-    0x3b,
-    0xa9,
-    0x0e,
-    0x0e,
-    0x5d,
-    0xd2,
-    0x44,
-    0xe3,
-    0x76,
-    0x14,
-    0x5c,
-    0xab,
-    0x4b,
-    0x60,
-    0xb9,
-    0xae,
-    0x08,
-    0xb2,
-    0x91,
-    0x59,
-    0xe6,
-    0x65,
-    0x08,
-    0x46,
-    0xbd,
-    0xd3,
-    0x7a,
-    0x1f,
-    0x44,
-    0xcb,
-    0x80,
-    0x51,
-    0x19,
-    0x35,
-    0x46,
-    0x9f,
-    0xd8,
-    0x04,
-    0x70,
-    0xec,
-    0xab,
-    0x7e,
-    0x6f,
-    0x03,
-    0x2c,
-    0x55,
-    0x48,
-    0x53,
-    0x06,
-    0x4c,
-    0xe1,
-    0xcf,
-    0x74,
-    0x6b,
-    0xaf,
-    0x25,
-    0x53,
-    0x53,
-    0xd3,
-    0x56,
-    0xf8,
-    0xb2,
-    0x59,
-    0x1b,
-    0xa6,
-    0x66,
-    0x48,
-    0x12,
-    0x92,
-    0x54,
-    0xf8,
-    0x15,
-    0xda,
-    0xdf,
-    0x81,
-    0xa0,
-    0x49,
-    0x79,
-    0xc3,
-    0x5c,
-    0x14,
-    0xdf,
-    0x20,
-    0xc0,
-    0xf6,
-    0xf6,
-    0x48,
-    0x50,
-    0xb0,
-    0x1c,
-    0x55,
-    0x09,
-    0x9c,
-    0xd2,
-    0xa5,
-    0x93,
-    0x78,
-    0xa6,
-    0xe1,
-    0x91,
-    0x1c,
-    0x4a,
-    0xe9,
-    0x83,
-    0xa0,
-    0x89,
-    0xd1,
-    0x97,
-    0x22,
-    0xca,
-    0x8d,
-    0xda,
-    0xf6,
-    0x68,
-    0x32,
-    0x70,
-    0x52,
-    0x74,
-    0xc0,
-    0xba,
-    0x57,
-    0xfb,
-    0xed,
-    0x70,
-    0xa7,
-    0x79,
-    0x3c,
-    0xf3,
-    0xab,
-    0x9d,
-    0x77,
-    0x1e,
-    0xc0,
-    0x91,
-    0x3d,
-    0xa7,
-    0xd7,
-    0xae,
-    0x60,
-    0x30,
-    0xf9,
-    0x60,
-    0x6c,
-    0xa4,
-    0x7d,
-    0x8b,
-    0x49,
-    0xce,
-    0x67,
-    0xdb,
-    0x23,
-    0xdb,
-    0x0e,
-    0xd6,
-    0x5a,
-    0x8f,
-    0xe0,
-    0x8c,
-    0xc7,
-    0x95,
-    0x97,
-    0x81,
-    0x03,
-    0x14,
-    0x36,
-    0xcc,
-    0xbe,
-    0x29,
-    0xee,
-    0xdd,
-    0x8e,
-    0x4f,
-    0xb0,
-    0xab,
-    0x83,
-    0xd0,
-    0x3e,
-    0x89,
-    0x2e,
-    0xc7,
-    0x8f,
-    0xdf,
-    0xda,
-    0x9a,
-    0xc5,
-    0x19,
-    0x53,
-    0x1c,
-    0xfb,
-    0xed,
-    0x26,
-    0x06,
-    0xd1,
-    0x21,
-    0x4c,
-    0x31,
-    0x27,
-    0x08,
-    0xd7,
-    0x82,
-    0x9c,
-    0x3d,
-    0x94,
-    0xc0,
-    0x4f,
-    0x67,
-    0x22,
-    0x9b,
-    0xa3,
-    0xd3,
-    0xdf,
-    0xdd,
-    0xbf,
-    0xd5,
-    0xbf,
-    0x24,
-    0xa1,
-    0x06,
-    0x64,
-    0x78,
-    0xbf,
-    0xa4,
-    0x79,
-    0xba,
-    0xef,
-    0x18,
-    0xef,
-    0xd0,
-    0x38,
-    0xc3,
-    0xf7,
-    0x68,
-    0xc1,
-    0xf8,
-    0xba,
-    0x9c,
-    0x52,
-    0x0b,
-    0xd2,
-    0xae,
-    0x71,
-    0x1e,
-    0xf6,
-    0x8b,
-    0x80,
-    0x86,
-    0x8a,
-    0x96,
-    0x36,
-    0xf3,
-    0x55,
-    0x88,
-    0x0c,
-    0xc6,
-    0x7f,
-    0x2b,
-    0x95,
-    0xe0,
-    0x9c,
-    0x80,
-    0xa7,
-    0x78,
-    0xf5,
-    0xa8,
-    0x37,
-    0xe5,
-    0x85,
-    0xc4,
-    0xcb,
-    0xac,
-    0x2c,
-    0x8e,
-    0x60,
-    0x4d,
-    0x29,
-    0x30,
-    0xa9,
-    0x78,
-    0xcd,
-    0xbb,
-    0x1d,
-    0x63,
-    0xee,
-    0xa6,
-    0x7d,
-    0x62,
-    0xfe,
-    0xbe,
-    0xa1,
-    0x9e,
-    0xec,
-    0x2a,
-    0xe9,
-    0x06,
-    0x5d,
-    0x09,
-    0xa3,
-    0x1a,
-    0x59,
-    0x89,
-    0x3f,
-    0x2d,
-    0xc7,
-    0x27,
-    0x2c,
-    0x73,
-    0x1f,
-    0x41,
-    0x80,
-    0xa0,
-    0x39,
-    0xc7,
-    0x0f,
-    0x87,
-    0xf6,
-    0x5d,
-    0xda,
-    0xe4,
-    0xc1,
-    0x01,
-    0xf4,
-    0x9d,
-    0x81,
-    0xa5,
-    0x6c,
-    0x5d,
-    0x19,
-    0x16,
-    0xc3,
-    0x21,
-    0x19,
-    0x64,
-    0xe7,
-    0x69,
-    0x32,
-    0x25,
-    0x77,
-    0xfa,
-    0xad,
-    0xfd,
-    0x2b,
-    0x61,
-    0xd4,
-    0x1e,
-    0x9d,
-    0x19,
-    0x3c,
-    0x3d,
-    0x39,
-    0xec,
-    0x5c,
-    0x9c,
-    0x97,
-    0x22,
-    0x64,
-    0x8b,
-    0x44,
-    0x37,
-    0x68,
-    0xf4,
-    0xdb,
-    0x5b,
-    0x7a,
-    0x5a,
-    0x2e,
-    0xb0,
-    0x9e,
-    0x90,
-    0xa7,
-    0xeb,
-    0x51,
-    0x2a,
-    0x72,
-    0x8e,
-    0x2d,
-    0xaa,
-    0xbf,
-    0x16,
-    0xa4,
-    0x33,
-    0x71,
-    0x1b,
-    0x29,
-    0xea,
-    0x8b,
-    0xb7,
-    0x10,
-    0x9e,
-    0x33,
-    0x89,
-    0xcb,
-    0x0e,
-    0xda,
-    0xbc,
-    0x7d,
-    0xc6,
-    0x67,
-    0xe4,
-    0xf4,
-    0x66,
-    0x66,
-    0xbc,
-    0x66,
-    0xd2,
-    0x0a,
-    0xff,
-    0xae,
-    0xfa,
-    0x50,
-    0x46,
-    0xcb,
-    0xbf,
-    0x00,
-    0xee,
-    0x08,
   ]);
 }
